Ascii-IIimport java.util.*; public class Ascii {     public static void main(String args[])     {         //int a;         Scanner sc=new Scanner(System.in);         int a=sc.nextInt();         System.out.println((char)a);      } }import java.util.*; public class Ascii {     public static void main(String args[])     {         //int a;         Scanner sc=new Scanner(System.in);         int a=sc.nextInt();         System.out.println((char)a);      } }
ASCII - American Standard Code for Information Interchange. Write a program to get a character as input and print it's ASCII value. When a character is entered, the character itself  stored.  And when we displayed that value using "int()"  type casting , the entered character is displayed as Ascii value. NOTE: Bold letters or words are Input and rest are output statements